refactor: onboarding을 profile 도메인에서 auth/signup으로 이동
