+++ features/auth/signup/index.ts
@@ -1,0 +1,2 @@
+export { useOnboarding } from './mutations';
+export type { OnboardingRequest } from './types';
