+++ features/auth/signup/mutations.ts
@@ -1,0 +1,8 @@
+import { useMutation } from '@tanstack/react-query';
+
+import { createOnboarding } from './fetchers';
+
+export const useOnboarding = () =>
+  useMutation({
+    mutationFn: createOnboarding,
+  });
