+++ features/auth/signup/types.ts
@@ -1,0 +1,5 @@
+export type OnboardingRequest = {
+  realName: string;
+  bio: string;
+  profileImageUrl?: string | null;
+};
