+++ features/profile/api/mutations.ts
@@ -1,12 +1,7 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 
-import { createOnboarding, updateProfile } from './fetchers';
+import { updateProfile } from './fetchers';
 import { profileKeys } from './queries';
-
-export const useOnboarding = () =>
-  useMutation({
-    mutationFn: createOnboarding,
-  });
 
 export const useUpdateProfile = () => {
   const queryClient = useQueryClient();
