+++ features/profile/index.ts
@@ -1,3 +1,3 @@
-export { useOnboarding, useUpdateProfile } from './api/mutations';
+export { useUpdateProfile } from './api/mutations';
 export { useProfile } from './api/queries';
-export type { OnboardingRequest, Profile, UpdateProfileRequest } from './types/profile';
+export type { Profile, UpdateProfileRequest } from './types/profile';
